snake action creators to own file
